refactor(simple_deep_agent): rename types.py to agent_types.py
The module shadowed the stdlib `types` module, which prevented mypy from
running in path mode (`mypy .`). Renaming unblocks full-tree type checking.
