+++ lib/workflows/simple_deep_agent/manifest_base.py
@@ -20,7 +20,7 @@ from lib.workflows.simple_deep_agent.state import (
     SimpleDeepAgentConfig,
     SimpleDeepAgentState,
 )
-from lib.workflows.simple_deep_agent.types import issues_from_agent_result
+from lib.workflows.simple_deep_agent.agent_types import issues_from_agent_result
 
 if TYPE_CHECKING:
     from lib.workflows.workflow_types import WorkflowState
